% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_models_inaugural.R
\docType{data}
\name{lda_models_inaugural}
\alias{lda_models_inaugural}
\title{A set of pre-run LDA models}
\format{An object of class \code{list} of length 14.}
\usage{
lda_models_inaugural
}
\description{
Few LDA models with different specifications run over the US presidential
inaugural address corpus available from \link{quanteda-package}. 
Information about the corpus can be found here 
\code{\link[quanteda]{data_corpus_inaugural}}.
}
\details{
A list object containing pre-run LDA models with 
k = 2, 5, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100, 150, 200.
}
\keyword{data}
