% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_topic_old_par.R
\name{optimal_topic_old_par}
\alias{optimal_topic_old_par}
\title{Find the optimal number of topics from a pool of LDA models}
\usage{
optimal_topic_old_par(
  lda_models,
  weighted_dfm,
  q = 0.8,
  alpha = 0.05,
  ncores = NULL,
  do_plot = TRUE,
  convert = NULL
)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as computed by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{weighted_dfm}{A weighted \code{\link[quanteda]{dfm}} containing word proportions.
It is recommended that \code{weighted_dfm} has element names consistent with the ones detected
by \code{\link[topicmodels]{LDA}}. See 'Details'.}

\item{q}{Set a cutoff for important words as the quantile of the expected
cumulative probability of word weights. Default to 0.80, meaning that the 
function reaches 80\% of the distribution mass and leaves out the remaining
20\%.}

\item{alpha}{The confidence level of test acceptance. Default to 0.05. 
See 'Details'.}

\item{ncores}{Control the number of cores to use through \code{\link[parallel]{detectCores}}.}

\item{do_plot}{Plot the chi-square statistic as a function of the number of 
topics. Default to \code{TRUE}.}

\item{convert}{Target convertion format. This version of \code{OpTop} supports
\code{\link[base]{data.frame}} and \code{\link[tibble]{tibble}}.
Default to \code{NULL} which returns a \code{\link[data.table]{data.table}}.}
}
\value{
A \code{data.table} containing the following columns:


\item{\code{topic}}{An integer giving the number of topics.}
\item{\code{OpTop}}{A numeric giving the standardized chi-square.}
\item{\code{pval}}{A numeric giving the p-value of the test.}
}
\description{
Implements a fast chi-square like test to detect the number of topics
that best describes the corpus estimated via Latent Dirichelet Allocation.
}
\details{
The function implements a Pearson chi-square statistic that exploits
the assumption that the distribution of words is multinomial. The test studies
the stability of a K-topic model which fully characterizes the corpus if the
observed and estimated word vectors are statistically indistinct.

To ensure a complete matching between the set of LDA models specified 
through \code{lda_models}, we strongly recommend that the corresponding \code{weighted_dfm} 
has specific element names indicating the original names of the documents as defined
in the \code{\link[quanteda]{corpus}}. These element names can be extracted with 
\code{\link[quanteda]{docid}}\code{(weighted_dfm)}. 
If, for any reason, the function \code{\link[topicmodels]{LDA}} fails 
to estimate the requested \code{k} topics over certain documents, then \code{optimal_topic} 
takes care of that by ensuring that there is a perfect match between the documents found in 
\code{weighted_dfm} and the ones contained in \code{lda_models}. 
If \code{weighted_dfm} does not contain any meaningful name to be matched with \code{lda_models}, 
for instance if the whole vector is full of \code{FALSE}, then \code{optimal_topic} stops 
with an error because most likely there is something wrong. If \code{optimal_topic} finds
few documents that are not present in \code{lda_models}, then it removes them from the input
\code{weighted_dfm} in order to achieve a perfect match. 
 
The parameter \code{alpha} controls the confidence of the chi-square test. The
optimal model is selected the first time the chi-square statistic reaches
a p-value equal to \code{alpha}. In the event that the chi-square statistic
fails to reach \code{alpha}, the minimum chi-square statistic
is selected. A higher \code{alpha} resolves in selecting a model with less 
topics. You can force the algorithm to find the minimum chi-square statistic
by setting \code{alpha} equal to zero.
}
\examples{
\dontrun{
# Compute word proportions from a corpus objects
test1 <- optimal_topic( lda_models = lda_list,
                        weighted_dfm = weighted_dfm,
                        q = 0.80,
                        alpha = 0.05 )
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}

Craig M. Lewis \email{craig.lewis@owen.vanderbilt.edu}
}
