% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpTop.R
\docType{package}
\name{OpTop-package}
\alias{OpTop-package}
\alias{OpTop}
\title{OpTop: detect the optimal number of topics from a pool of LDA models}
\description{
OpTop is an \code{R} package that implements the testing approach described in
the paper A Statistical Approach for Optimal Topic Model Identification
by Lewis and Grossetti (2019).
}
\details{
OpTop introduces a set of parametric tests to identify the optimal number
of topics in a collection of LDA models. OpTop also includes several
tests to explore topic stability and redundancy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/contefranz/OpTop}
  \item Report bugs at \url{https://github.com/contefranz/OpTop/issues}
}

}
\author{
\strong{Maintainer}: Francesco Grossetti \email{francesco.grosssetti@unibocconi.it} (\href{https://orcid.org/0000-0002-5130-7745}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Craig Lewis \email{craig.lewis@owen.vanderbilt.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item Mattia Tamellini \email{mattia.tamellini@gmail.com} [contributor]
}

}
\keyword{internal}
