% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_match.R
\name{topic_match}
\alias{topic_match}
\title{Define informative and uninformative components}
\usage{
topic_match(lda_models, optimal_model, var_correction = TRUE)
}
\arguments{
\item{lda_models}{A list of \code{topicmodels::LDA} objects (VEM), ordered by
increasing number of topics. The grid should span the candidate values of \code{K}.}

\item{optimal_model}{Integer giving the number of topics of the optimal model;
must match one entry in \code{lda_models}.}

\item{var_correction}{Use the unbiased estimator of the co-variance for
i.i.d. observations by applying n - 1 in the denominator. Default is \code{TRUE}.}
}
\value{
A named list with the informative and uninformative components given
as matrices.
}
\description{
For above-optimal topics, detect and extract informative and uninformative
components in terms of cosine similarities with the optimal topic model.
}
\examples{
\dontrun{
out3 <- topic_match( lda_models = lda_list, 
                     optimal_model = test1, 
                     var_correction = TRUE )
}
}
\keyword{internal}
