% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_proportions.R
\name{word_proportions}
\alias{word_proportions}
\title{Compute word proportions from a corpus object}
\usage{
word_proportions(corpus, remove_document = FALSE, language = "en",
  source = "snowball")
}
\arguments{
\item{corpus}{A \code{\link[quanteda]{corpus}} object as defined in
\code{quanteda}.}

\item{remove_document}{Remove the \code{document} identification inherited
from \code{\link[quanteda]{corpus}}. Default to \code{TRUE}.}

\item{language}{A character specifying the language of stopwords according
to ISO 639-1 code. It mimics the behavior of \code{\link[stopwords]{stopwords}}.
Default to \code{"en"} for English.}

\item{source}{A character specifying the stopwords source. It can be either
\code{"snowball"}, \code{"snowball-iso"}, \code{"misc"}, or \code{"smart"}.
Default to \code{"snowball"} for consistency with
\code{\link[stopwords]{stopwords}}.}
}
\value{
A \code{data.table} with the following columns:
\item{\code{id_doc}}{A sequential integer giving the identification of documents.}
\item{\code{id_word}}{A sequential integer giving the identification of words.}
\item{\code{word}}{A character identifying the word.}
\item{\code{word_count}}{An integer giving the word count.}
\item{\code{word_prop}}{A numeric giving the word proportion.}
}
\description{
Fast routine that computes word proportions from a
\code{\link[quanteda]{corpus}}.
}
\examples{
\dontrun{
# Compute word proportions from a corpus objects
word_proportions = word_proportions( corpus = data_corpus_inaugural,
                                     remove_document = TRUE,
                                     language = "en",
                                     source = "snowball" )
}
}
\references{
Stopwords ISO: \url{https://github.com/stopwords-iso/stopwords-iso}.\cr

Full list of stopwords:
\url{https://github.com/stopwords-iso/stopwords-iso/blob/master/CREDITS.md}.
}
\seealso{
\code{\link[quanteda]{corpus}} \code{\link[data.table]{data.table}}
\code{\link[stopwords]{stopwords}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}.
}
