% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_document_stability.R
\name{agg_document_stability}
\alias{agg_document_stability}
\title{Compute aggregate document stability}
\usage{
agg_document_stability(lda_models, word_proportions, optimal_model,
  threshold = 0.00075, convert = FALSE)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as computed by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{word_proportions}{A \code{data.table} giving the word proportions in a corpus
as computed by \code{\link[OpTop]{word_proportions}}.}

\item{optimal_model}{The optimal topic model. It can be either an integer giving
the number of topics for the optimal topic model or a
\code{\link[data.table]{data.table}}/\code{\link[base]{data.frame}} as
returned by \code{\link[OpTop]{optimal_topic}}. In the latter case, the
function automatically finds the optimal model. This argument can also be
a \code{\link[topicmodels]{LDA-class}} object of type \code{LDA_VEM} as 
returned by \code{\link[topicmodels]{LDA}}.}

\item{threshold}{Set a cutoff for the relative importance of words.
Default to 0.00075.}

\item{convert}{Convert the output to a \code{data.frame}.
Default to \code{FALSE}.}
}
\description{
Detects informative and uninformative components to compute aggregate document
stability.
}
\examples{
\dontrun{
test4 <- agg_topic_stability( lda_models = lda_list,
                              best_k = test1,
                              best_match = BestMatch,
                              least_match = LeastMatch )
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}.
}
