% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_document_stability.R
\name{agg_document_stability}
\alias{agg_document_stability}
\title{Compute aggregate document stability and F-test}
\usage{
agg_document_stability(
  lda_models,
  weighted_dfm,
  optimal_model,
  q = 0.8,
  alpha = 0.05,
  smoothed = TRUE,
  do_plot = TRUE,
  convert = NULL
)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as computed by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{weighted_dfm}{A weighted \code{\link[quanteda]{dfm}} containing word proportions.
It is recommended that \code{weighted_dfm} has the \code{\link[quanteda]{docvar}} 
"doc_id" with original document names. See ?\code{\link[OpTop]{optimal_topic}} for more details.}

\item{optimal_model}{The optimal topic model. It can be either an integer giving
the number of topics for the optimal topic model or a
\code{\link[data.table]{data.table}}/\code{\link[base]{data.frame}} as
returned by \code{\link[OpTop]{optimal_topic}}. In the latter case, the
function automatically finds the optimal model. This argument can also be
a \code{\link[topicmodels]{LDA-class}} object of type \code{LDA_VEM} as 
returned by \code{\link[topicmodels]{LDA}}.}

\item{q}{Set a cutoff for important words as the quantile of the expected
cumulative probability of word weights. Default to 0.80, meaning that the 
function reaches 80\% of the distribution mass and leaves out the remaining
20\%.}

\item{alpha}{Alpha level to identify informative words from the Cumulative
Distribution Function over the cosine similarities in the Topic Word Weights
matrix. Default to 0.05.}

\item{smoothed}{A logical to control whether the test is performed on each 
document for each LDA model or on the smoothed chi-square statistic. 
This is the aggregated version which gives the overall behavior across all
documents in the corpus. Default is \code{TRUE}.}

\item{do_plot}{Plot the chi-square statistic as a function of the number of 
topics. Default to \code{TRUE}. The horizontal dot-dashed line represents
the significance level according to \code{alpha}.}

\item{convert}{Target convertion format. This version of \code{OpTop} supports
\code{\link[base]{data.frame}} and \code{\link[tibble]{tibble}}.
Default to \code{NULL} which returns a \code{\link[data.table]{data.table}}.}
}
\value{
A \code{data.table} containing the following columns:

* \code{topic} An integer giving the number of topics.
* \code{id_doc} An integer document id as given in the original corpus.
* \code{chisq_inform_std} A numeric giving the standardized chi-square statistic
for the informative component.
* \code{chisq_uninform_std} A numeric giving the standardized chi-square statistic
for the uninformative component.
* \code{pval_inform} A numeric giving the p-value of the chi-square test
over the informative component.
* \code{pval_uninform} A numeric giving the p-value of the chi-square test
over the uninformative component.
* \code{Fstat} A numeric giving the standardized F statistic
of the ratio \code{chisq_inform_std}/\code{chisq_uninform_std}.
* \code{pval_Fstat} A numeric giving the p-value of the F test.
}
\description{
Detects informative and uninformative components to compute aggregate document
stability. Performs a chi-square test to evaluate document stability, 
Also, computes a F-test to further evaluate deviation from optimal model.
}
\examples{
\dontrun{
test4 <- agg_document_stability( lda_models = lda_list,
                                 weighted_dfm = weighted_dfm,
                                 smoothed = TRUE, do_plot = TRUE )
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}.

Craig M. Lewis \email{craig.lewis@owen.vanderbilt.edu}
}
