% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_topic_stability.R
\name{agg_topic_stability}
\alias{agg_topic_stability}
\title{Compute aggregate topic stability}
\usage{
agg_topic_stability(
  lda_models,
  optimal_model,
  q = 0.8,
  alpha = 0.05,
  smoothed = TRUE,
  do_plot = TRUE
)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as estimated by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{optimal_model}{A number corresponding to the optimal topic model.}

\item{q}{Set a cutoff for important words as the quantile of the expected
cumulative probability of word weights. Default to 0.80, meaning that the 
function reaches 80\% of the distribution mass and leaves out the remaining
20\%.}

\item{alpha}{Alpha level to identify informative words from the Cumulative
Distribution Function over the cosine similarities in the Topic Word Weights
matrix. Default to 0.05.}

\item{smoothed}{A logical to control whether the test is performed on each 
document for each LDA model or on the smoothed chi-square statistic. 
This is the aggregated version which gives the overall behavior across all
documents in the corpus. Default is \code{TRUE}.}

\item{do_plot}{Plot the chi-square statistic as a function of the number of 
topics. Default to \code{TRUE}. The horizontal dot-dashed line represents
the significance level according to \code{alpha}.}
}
\value{
A \code{data.table} containing the following columns:

\item{\code{topic}}{An integer giving the number of topics.}
\item{\code{id_doc}}{An integer document id as given in the original corpus.}
\item{\code{chisq_std}}{A numeric giving the standardized chi-square.}
\item{\code{pval}}{A numeric giving the p-value of the test.}
}
\description{
Compute aggregate topic and perform a chi-square test to evaluate topic
stability for above optimal models.
}
\examples{
\dontrun{
test4 <- agg_topic_stability( lda_models = lda_list,
                              optimal_model = test1)
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}

Craig M. Lewis \email{craig.lewis@owen.vanderbilt.edu}
}
