% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_topic_stability.R
\name{agg_topic_stability}
\alias{agg_topic_stability}
\title{Compute aggregate topic stability}
\usage{
agg_topic_stability(lda_models, optimal_model, best_match, least_match,
  threshold = 0.00075, do_plot = TRUE, convert = FALSE)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as computed by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{optimal_model}{The optimal topic model. It can be either an integer giving
the number of topics for the optimal topic model or a
\code{\link[data.table]{data.table}}/\code{\link[base]{data.frame}} as
returned by \code{\link[OpTop]{optimal_topic}}. In the latter case, the
function automatically finds the optimal model. This argument can also be
a \code{\link[topicmodels]{LDA-class}} object of type \code{LDA_VEM} as 
returned by \code{\link[topicmodels]{LDA}}.}

\item{best_match}{A data.table as computed by \code{\link[OpTop]{topic_match}}.
This contains the models with highest cosine similarity.}

\item{least_match}{A data.table as computed by \code{\link[OpTop]{topic_match}}.
This contains the models with lowest cosine similarity.}

\item{threshold}{Set a cutoff for the relative importance of words.
Default to 0.00075.}

\item{do_plot}{Plot the chi-square statistic as a function of the number of 
topics. Default is \code{TRUE}.}

\item{convert}{Convert the output to a \code{data.frame}.
Default to \code{FALSE}.}
}
\description{
Detects informative and uninformative components to compute aggregate topic
stability.
}
\examples{
\dontrun{
test4 <- agg_topic_stability( lda_models = lda_list,
                              optimal_model = test1,
                              best_match = BestMatch,
                              least_match = LeastMatch )
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}

Craig M. Lewis \email{craig.lewis@owen.vanderbilt.edu}
}
