% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topic_models.R
\name{get_topic_models}
\alias{get_topic_models}
\title{Get the list of topic models from a specified environment}
\usage{
get_topic_models(object_type, pattern, envir = globalenv())
}
\arguments{
\item{object_type}{A character specifying the class of objects to extract.
Default to \code{LDA_VEM} as given by the virtual class
\code{\link[topicmodels]{TopicModel-class}}.}

\item{pattern}{an optional \link[base]{regular expression}.  Only names
    matching \code{pattern} are returned.  \code{\link{glob2rx}} can be
    used to convert wildcard patterns to regular expressions.}

\item{envir}{The environment where to search for topic models. Default to
\code{\link[base]{globalenv}}.}
}
\value{
A named list containing the objects of specified class.
}
\description{
Easily get the list of topic models from a specific class. This is a handy
feature to facilitate the identification of only the objects the user wants
to process.
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}.
}
