% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_topic.R
\name{optimal_topic}
\alias{optimal_topic}
\title{Find the optimal number of topics from a pool of LDA models}
\usage{
optimal_topic(lda_models, word_proportions, threshold = 0.00075,
  alpha = 0.01, q_type = 5L, do_plot = TRUE, convert = FALSE)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as computed by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{word_proportions}{A \code{data.table} giving the word proportions in a corpus
as computed by \code{\link[OpTop]{word_proportions}}.}

\item{threshold}{Set a cutoff for the relative importance of words.
Default to 0.00075.}

\item{alpha}{Probability at which compute the quantiles of the chi-square test
and adjust its values. Default to 0.01.}

\item{q_type}{Select the quantile algorithm as in \code{\link[stats]{quantile}}.
Default to 5 for consistency with Matlab. In later releases, it will be
replaced with 7, which is the \code{R} default.}

\item{do_plot}{Plot the chi-square statistic as a function of the number of 
topics. Default is \code{TRUE}.}

\item{convert}{Convert the output to a \code{data.frame}.
Default to \code{FALSE}.}
}
\value{
A \code{data.table} containing the following columns:

\item{\code{topic}}{An integer giving the number of topics.}
\item{\code{chi_sum}}{A numeric giving the overall sum of the chi-square.}
\item{\code{cut}}{An integer giving the row position of minimum word
importance.}
\item{\code{chi_std}}{A numeric giving the standardized chi-square.}
\item{\code{chi_stdw}}{A numeric giving the standardized and weighted chi-square.}
}
\description{
Implements a fast chi-square like test to detect the number of topics
that best describes the corpus.
}
\details{
The function implements a Pearson chi-square statistic that exploits
the assumption that the distribution of words is multinomial. The test studies
the stability of a K-topic model which fully characterizes the corpus if the
observed and estimated word vectors are statistically indistinct.
}
\examples{
\dontrun{
# Compute word proportions from a corpus objects
word_proportions <- word_proportions( corpus = data_corpus_inaugural,
                                     remove_document = TRUE,
                                     language = "en",
                                     source = "snowball" )

test1 <- optimal_topic( lda_models = lda_list,
                       word_proportions = word_proportions,
                       threshold = 0.00075,
                       alpha = 0.01,
                       q_type = 5 )
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}

Craig M. Lewis \email{craig.lewis@owen.vanderbilt.edu}
}
