% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_topic.R
\name{optimal_topic}
\alias{optimal_topic}
\title{Find the optimal number of topics from a pool of LDA models}
\usage{
optimal_topic(lda_models, word_proportions, q = 0.8, alpha = 0.05,
  do_plot = TRUE, convert = NULL)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as computed by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{word_proportions}{A \code{\link[data.table]{data.table}} giving the 
word proportions in a corpus
as computed by \code{\link[OpTop]{word_proportions}}.}

\item{q}{Set a cutoff for important words as the quantile of the expected
cumulative probability of word weights. Default to 0.80, meaning that the 
function reaches 80\% of the distribution mass and leaves out the remaining
20\%.}

\item{alpha}{The confidence level of test acceptance. Default to 0.05. 
See 'Details'.}

\item{do_plot}{Plot the chi-square statistic as a function of the number of 
topics. Default to \code{TRUE}.}

\item{convert}{Target convertion format. This version of \code{OpTop} supports
\code{\link[base]{data.frame}} and \code{\link[tibble]{tibble}}.
Default to \code{NULL} which returns a \code{\link[data.table]{data.table}}.}
}
\value{
A \code{data.table} containing the following columns:

\item{\code{topic}}{An integer giving the number of topics.}
\item{\code{chisq_std}}{A numeric giving the standardized chi-square.}
\item{\code{pval}}{A numeric giving the p-value of the test.}
}
\description{
Implements a fast chi-square like test to detect the number of topics
that best describes the corpus estimated via Latent Dirichelet Allocation.
}
\details{
The function implements a Pearson chi-square statistic that exploits
the assumption that the distribution of words is multinomial. The test studies
the stability of a K-topic model which fully characterizes the corpus if the
observed and estimated word vectors are statistically indistinct.

The parameter \code{alpha} controls the confidence of the chi-square test. The
optimal model is selected the first time the chi-square statistic reaches
a p-value equal to \code{alpha}. In the event that the chi-square statistic
fails to reach \code{alpha}, the minimum chi-square statistic
is selected. A higher \code{alpha} resolves in selecting a model with less 
topics. You can force the algorithm to find the minimum chi-square statistic
by setting \code{alpha} equal to zero.
}
\examples{
\dontrun{
# Compute word proportions from a corpus objects
word_proportions <- word_proportions( corpus = data_corpus_inaugural,
                                      remove_document = TRUE,
                                      language = "en",
                                      source = "snowball" )

test1 <- optimal_topic( lda_models = lda_list,
                        word_proportions = word_proportions,
                        q = 0.80,
                        alpha = 0.05 )
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}

Craig M. Lewis \email{craig.lewis@owen.vanderbilt.edu}
}
