% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dfm.R
\name{sim_dfm}
\alias{sim_dfm}
\title{Simulate a document-feature-matrix from a LDA specification}
\usage{
sim_dfm(DTW, TWW, doc_lenght, alpha = NULL, seed = NULL)
}
\arguments{
\item{DTW}{A matrix or data.frame with Document-Topic-Weights.}

\item{TWW}{A matrix or data.frame with Topic-Word-Weights.}

\item{doc_lenght}{A vector containing the desired document length as total number of word counts.}

\item{alpha}{Parameter of the Dirichlet distribution for topics over documents.}

\item{seed}{Input to \code{set.seed}.}
}
\value{
A \code{\link[quanteda]{dfm}} object.
}
\description{
Simulate a document-feature-matrix from estimated DTW and TWW. The number of topics as well as 
the number of documents are inferred from the LDA parameters.
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[quanteda]{dfm}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}

Craig M. Lewis \email{craig.lewis@owen.vanderbilt.edu}
}
