% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_match.R
\name{topic_match}
\alias{topic_match}
\title{Define informative and uninformative components}
\usage{
topic_match(lda_models, best_k, var_correction = TRUE)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as computed by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{best_k}{The optical topic model. It can be either an integer giving
the number of topics for the optimal topic model or a
\code{\link[data.table]{data.table}}/\code{\link[base]{data.frame}} as
returned by \code{\link[OpTop]{optimal_topic}}. In the latter case, the
function automatically finds the optimal model.}

\item{var_correction}{Use the unbiased estimator of the co-variance for 
i.i.d. observations by applying n - 1 in the denominator. Default is \code{TRUE}.}
}
\value{
A named list with the informative and uninformative components given
as matrices.
}
\description{
For above-optimal topics, detect and extract informative and uninformative
components in terms of cosine similarities with the optimal topic model.
}
\examples{
\dontrun{
out3 <- topic_match( lda_models = lda_list, 
                     best_k = test1, 
                     var_correction = TRUE )
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}.
}
