% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_match.R
\name{topic_match}
\alias{topic_match}
\title{Define informative and uninformative components}
\usage{
topic_match(lda_models, optimal_model, var_correction = TRUE)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as estimated by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{optimal_model}{The optimal topic model. It can be either an integer giving
the number of topics for the optimal topic model or a
\code{\link[data.table]{data.table}}/\code{\link[base]{data.frame}} as
returned by \code{\link[OpTop]{optimal_topic}}. In the latter case, the
function automatically finds the optimal model. This argument can also be
a \code{\link[topicmodels]{LDA-class}} object of type \code{LDA_VEM} as 
returned by \code{\link[topicmodels]{LDA}}.}

\item{var_correction}{Use the unbiased estimator of the co-variance for 
i.i.d. observations by applying n - 1 in the denominator. Default is \code{TRUE}.}
}
\value{
A named list with the informative and uninformative components given
as matrices.
}
\description{
For above-optimal topics, detect and extract informative and uninformative
components in terms of cosine similarities with the optimal topic model.
}
\examples{
\dontrun{
out3 <- topic_match( lda_models = lda_list, 
                     optimal_model = test1, 
                     var_correction = TRUE )
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}

Craig M. Lewis \email{craig.lewis@owen.vanderbilt.edu}
}
