% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_stability.R
\name{topic_stability}
\alias{topic_stability}
\title{Compute the stability of above optimum topics}
\usage{
topic_stability(lda_models, best_k, threshold = 0.00075, alpha = 0.05,
  compute_res = FALSE, convert = FALSE)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as computed by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{best_k}{The optical topic model. It can be either an integer giving
the number of topics for the optimal topic model or a
\code{\link[data.table]{data.table}}/\code{\link[base]{data.frame}} as
returned by \code{\link[OpTop]{optimal_topic}}. In the latter case, the
function automatically finds the optimal model.}

\item{threshold}{Set a cutoff for the relative importance of words.
Default to 0.00075.}

\item{alpha}{Alpha level to identify informative words. Default to 0.05.}

\item{compute_res}{Determines whether or not the function has to compute
cosine similarities for unmatched topic.}

\item{convert}{Convert the output to a \code{data.frame}.
Default to \code{FALSE}.}
}
\description{
Implements a fast chi-square like test to evaluate the stability of redundant
topics.
}
\examples{
\dontrun{
test2 <- topic_stability( lda_models = lda_list,
                          best_k = test1,
                          threshold = 0.00075,
                          alpha = 0.05 )
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}.
}
