% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_stability.R
\name{topic_stability}
\alias{topic_stability}
\title{Compute topic stability for over-optimal topic specifications}
\usage{
topic_stability(
  lda_models,
  optimal_model,
  q = 0.8,
  alpha = 0.05,
  do_plot = TRUE,
  convert = NULL
)
}
\arguments{
\item{lda_models}{A list of ordered LDA models as computed by
\code{\link[topicmodels]{LDA}}. The LDA models must be in ascending order
according to the number of topics.}

\item{optimal_model}{The optimal topic model. It can be either an integer giving
the number of topics for the optimal topic model or a
\code{\link[data.table]{data.table}}/\code{\link[base]{data.frame}} as
returned by \code{\link[OpTop]{optimal_topic}}. In the latter case, the
function automatically finds the optimal model. This argument can also be
a \code{\link[topicmodels]{LDA-class}} object of type \code{LDA_VEM} as 
returned by \code{\link[topicmodels]{LDA}}.}

\item{q}{Set a cutoff for important words as the quantile of the expected
cumulative probability of word weights. Default to 0.80, meaning that the 
function reaches 80\% of the distribution mass and leaves out the remaining
20\%.}

\item{alpha}{Alpha level to identify informative words from the Cumulative
Distribution Function over the cosine similarities in the Topic Word Weights
matrix. Default to 0.05.}

\item{do_plot}{Plot the chi-square statistic as a function of the number of 
topics. Default to \code{TRUE}. The horizontal dot-dashed line represents
the significance level according to \code{alpha}.}

\item{convert}{Target convertion format. This version of \code{OpTop} supports
\code{\link[base]{data.frame}} and \code{\link[tibble]{tibble}}.
Default to \code{NULL} which returns a \code{\link[data.table]{data.table}}.}
}
\value{
A `data.table` containing the following columns:

\item{\code{topic}}{An integer giving the number of topics.}
\item{\code{df}}{An integer giving the degrees of freedom.}
\item{\code{chisq}}{A numeric giving the chi-square statistic.}
}
\description{
Implements fast chi-square like test to evaluate the stability of redundant
topics.
}
\details{
This function implements Test 3 as defined in 
Lewis and Grossetti (2019). Test 3 evaluates the aggregated 
stability of over-optimal topic specifications by summing each 
point-wise contribution. See 'Value' to understand how \code{topic_stability} 
returns the results.
}
\examples{
\dontrun{
test2 <- topic_stability( lda_models = lda_list,
                          optimal_model = out,
                          q = 0.00075,
                          alpha = 0.05 )
}
}
\references{
Lewis, C. and Grossetti, F. (2019 - forthcoming):\cr
A Statistical Approach for Optimal Topic Model Identification.
}
\seealso{
\code{\link[topicmodels]{LDA}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}

Craig M. Lewis \email{craig.lewis@owen.vanderbilt.edu}
}
