% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_proportions.R
\name{word_proportions}
\alias{word_proportions}
\title{Compute word proportions from a corpus object}
\usage{
word_proportions(corpus, remove_document = FALSE)
}
\arguments{
\item{corpus}{A \code{\link[quanteda]{corpus}} object as defined in
\code{quanteda}.}

\item{remove_document}{Remove the \code{document} identification inherited
from \code{\link[quanteda]{corpus}}. Default to \code{TRUE}.}
}
\value{
A \code{data.table} with the following columns:
\item{\code{id_doc}}{A sequential integer giving the identification of documents.}
\item{\code{id_word}}{A sequential integer giving the identification of words.}
\item{\code{word}}{A character identifying the word.}
\item{\code{word_count}}{An integer giving the word count.}
\item{\code{word_prop}}{A numeric giving the word proportion.}
}
\description{
Fast routine that computes word proportions from a
\code{\link[quanteda]{corpus}}.
}
\examples{
\dontrun{
# Compute word proportions from a corpus objects
word_proportions = word_proportions( data_corpus_inaugural, remove_document = TRUE )
}
}
\seealso{
\code{\link[quanteda]{corpus}} \code{\link[data.table]{data.table}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}.
}
