% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_proportions.R
\name{word_proportions}
\alias{word_proportions}
\title{Compute word proportions from a corpus or a dfm object}
\usage{
word_proportions(x, remove_document = FALSE, remove_nonASCII = TRUE, ...)
}
\arguments{
\item{x}{Either a \code{\link[quanteda]{corpus}} or 
\code{\link[quanteda]{dfm}} object as defined in \code{quanteda}.}

\item{remove_document}{Remove the \code{document} identification inherited
from \code{x}. Default to \code{FALSE} (See 'Details').}

\item{remove_nonASCII}{A logical to remove non-ASCII characters from the 
\code{x}. Default to \code{TRUE}.}

\item{...}{When \code{x} is a corpus, additional arguments passed 
to \code{\link[quanteda]{tokens}} and \code{\link[quanteda]{dfm}} to allow
for precision in tokens removal.}
}
\value{
A \code{data.table} with the following columns:
\item{\code{document}}{A character giving the original document identification
as inherited from the corpus or the dfm. This column is only retained when 
\code{remove_document = FALSE}.} 
\item{\code{id_doc}}{A sequential integer giving the document identification.}
\item{\code{id_word}}{A sequential integer giving the identification of words.}
\item{\code{word}}{A character identifying the word.}
\item{\code{word_count}}{An integer giving the word count.}
\item{\code{word_prop}}{A numeric giving the word proportion.}
}
\description{
Fast routine that computes word proportions from a
\code{\link[quanteda]{corpus}} or \code{\link[quanteda]{dfm}} object.
}
\details{
We recommend to keep \code{remove_document = FALSE} since
this triggers further controls before estimating the optimal model specification.
We refer the user to the inline help of \code{\link[OpTop]{optimal_topic}}.

\code{word_proportions} only applies preprocessing when \code{x} is a 
\code{\link[quanteda]{corpus}}. You can pass the usual parameters to \code{...}.
This includes the two parameters \code{remove_document} and \code{remove_nonASCII}.
If \code{x} is \code{\link[quanteda]{dfm}}, then the function directly
computes the word proportions as expected.
}
\examples{
\dontrun{
# Compute word proportions from a corpus object
word_proportions <- word_proportions( x = data_corpus_inaugural,
                                      remove = stopwords(), tolower = TRUE )
# Compute word proportions from a dfm object
word_proportions <- word_proportions( x = data_dfm_lbgexample )
}
}
\references{
Stopwords ISO: \url{https://github.com/stopwords-iso/stopwords-iso}.\cr

Full list of stopwords:
\url{https://github.com/stopwords-iso/stopwords-iso/blob/master/CREDITS.md}.
}
\seealso{
\code{\link[quanteda]{corpus}} \code{\link[data.table]{data.table}}
\code{\link[stopwords]{stopwords}}
}
\author{
Francesco Grossetti \email{francesco.grossetti@unibocconi.it}.
}
